refactor: optimize TCP packet handling and session structure - Added session ID to TCP sessions - Optimized packet handling logic by reducing local variables and using   unique_ptr - Refined previously placeholder packet processing logic - Prepared for future refactor to introduce packet-type-specific   structures

+++ matchserver/include/ServerTypes.h
@@ -1,14 +1,12 @@
 #ifndef SERVER_TYPES_H
 #define SERVER_TYPES_H
 
-using MatchId = uint32_t;
+using MatchId = uint16_t;
 using RequestId = uint32_t;
 using SessionId = uint32_t;
 using ServerTimeMs = uint32_t;
 
 constexpr uint16_t MAX_PACKET_SIZE{ 1024 };
 using RawBuffer = std::array<uint8_t, MAX_PACKET_SIZE>;
-
-
 
 #endif // !SERVER_TYPES_H
